clean(Wallet): moved privateKeyStringToKey to Helpers
